% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{circuits}
\alias{circuits}
\title{Circuits}
\format{
a data frame with the following columns:
\itemize{
\item circuitId (numeric)
\item circuitRef (character)
\item name (character)
\item location (character)
\item country (character)
\item lat (latitude, numeric)
\item lng (longditude, numeric)
\item alt (altitude, numeric)
\item length (km, numeric)
\item type (street, race or road, character)
\item direction (clockwise or anti-clockwise)
\item url (to wikipedia, character)
}
}
\source{
http://ergast.com/api/f1/circuits, with additions from wikipedia, others
}
\usage{
circuits
}
\description{
a data frame with information on all circuits used
in formula 1 history, from Ergast API
}
\keyword{datasets}
