% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{constructor_standings}
\alias{constructor_standings}
\title{Constructor Standings}
\format{
A data frame with the following columns:
\itemize{
\item constructorStandingsId (key, integer)
\item raceId (see \code{races}, integer)
\item constructorId (see \code{constructors}, integer)
\item points (numeric)
\item position (ranking at that point in the season, numeric)
\item positionText (ranking at that point in the season, character)
\item wins (number of wins that season, integer)
}
}
\source{
https://ergast.com/api/f1/standings
}
\usage{
constructor_standings
}
\description{
a list of drivers' standings at the end of every race
}
\keyword{datasets}
