% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{constructors}
\alias{constructors}
\title{Constructors}
\format{
a data frame with the following columns:
\itemize{
\item constructorId (unique Id, numeric)
\item constructorRef (a unique Id, character)
\item name (character)
\item nationality (character)
\item url (wikipedia url, character)
}
}
\source{
https://ergast.com/api/f1/constructors
}
\usage{
constructors
}
\description{
a list of constructors, nationalities, and urls
}
\keyword{datasets}
