% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{.combineData}
\alias{.combineData}
\title{Combine Data}
\usage{
.combineData(
  driverCrashEWMA = 0.05,
  carFailureEWMA = 0.05,
  tireFailureEWMA = 0.05,
  disqualifiedEWMA = 0.05,
  gridPositionCorEWMA = 0.2
)
}
\arguments{
\item{driverCrashEWMA}{How much one race impacts driver's moving average of crash rate}

\item{carFailureEWMA}{How much one race impacts constructor's moving average of failure rate}

\item{tireFailureEWMA}{How much one race impacts driver's moving average of tire failure rate}

\item{disqualifiedEWMA}{How much one race impacts driver's moving average of disqualified rate}

\item{gridPositionCorEWMA}{How much one race impacts driver's moving average of crash rate}
}
\value{
a tibble with tons of data
}
\description{
merge all of the data frames into one big one for use in modelling. EWMA parameters provided
help with determining the strength of the exponential weighted moving average for those parameters.
}
