% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{drivers}
\alias{drivers}
\title{Drivers}
\format{
a data frame with the following columns:
\itemize{
\item driverId (integer, unique)
\item driverRef (character, unique)
\item number (integer)
\item code (character)
\item forename (character)
\item surname (character)
\item dob (character, as YYYY-MM-DD)
\item nationality (character)
\item url (to wikipedia)
}
}
\source{
https://ergast.com/api/f1/drivers
}
\usage{
drivers
}
\description{
a list of drivers' names, numbers, codes, and id
}
\keyword{datasets}
