% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{pit_stops}
\alias{pit_stops}
\title{Pit Stops}
\format{
A data frame with the following columns:
\itemize{
\item raceId (see \code{races}, integer)
\item driverId (see \code{drivers}, integer)
\item stop (which stop number for that driver that race, integer)
\item lap (lap on which the driver came in to the pits, integer)
\item time (clock time of the pit entry, character)
\item duration (duration of the pit stop from pit entry to exit, numeric)
\item millisecond (duration fo the pit stop in milliseconds, integer)
}
}
\source{
https://ergast.com/api/f1/pit_stops
}
\usage{
pit_stops
}
\description{
a list of pit stops from 2011 onward
}
\keyword{datasets}
