% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{practices}
\alias{practices}
\title{Practices}
\format{
a data frame with the following columns:
\itemize{
\item position (finishing position in practice, integer)
\item driverNum (drivers' racing number (e.g. Hamilton = 44), integer)
\item driverName (drivers' name, character)
\item driverCar (drivers' car name, character)
\item time (fastest lap, character)
\item gap (gap to leader, character)
\item laps (number of laps, integer)
\item practiceNum (practice number, 1 through 4, integer)
\item raceId (race Id, see \code{races}, integer)
\item f1RaceId (race reference code, see \code{races}, character)
\item year (year, integer)
\item round (round number, integer)
\item driverId (driver reference code, see \code{drivers}, integer)
\item constructorId (constructor reference code, see \code{constructors}, integer)
}
}
\source{
f1.com
}
\usage{
practices
}
\description{
a list of practice results from 2000 onward
}
\keyword{datasets}
