% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{qualifying}
\alias{qualifying}
\title{Qualifying Results}
\format{
a data frame with the following columns:
\itemize{
\item qualifyId (key, integer)
\item raceId (see \code{races}, integer)
\item driverId (see \code{drivers}, integer)
\item constructorId (see \code{constructors}, integer)
\item number (driver number, integer)
\item position (finishing position, integer)
\item q1, q2, q3 (qualifying lap times, character)
}
}
\source{
https://ergast.com/api/f1/qualifying
}
\usage{
qualifying
}
\description{
a list of qualifying results by race
}
\keyword{datasets}
