% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{races}
\alias{races}
\title{Races}
\format{
a data frame with the following columns:
\itemize{
\item raceId (numeric)
\item year (numeric)
\item round (numeric)
\item circuitId (corresponding to a circuitId in circuits, numeric)
\item name (character)
\item date (character, as YYYY-MM-DD)
\item time (character, as ##:##:##)
\item url (character, to wikipedia)
\item weather (character, one of warm, wet, cloudy, dry, or unknown)
\item safetyCars (numeric)
\item safetyCarLaps (numeric)
}
}
\source{
https://ergast.com/api/f1
}
\usage{
races
}
\description{
a data frame containing race information
}
\keyword{datasets}
