% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{results}
\alias{results}
\title{Results}
\format{
a data frame with the following columns:
\itemize{
\item resultId, raceId, driverId, constructorId, number, grid, position (numeric)
\item positionText (character)
\item points, laps (numeric)
\item time, milliseconds, fastestLap, rank, fastestLapTime, fastestLapSpeed (character)
\item statusId (numeric)
\item status (character)
}
}
\source{
https://ergast.com/api/f1/results
}
\usage{
results
}
\description{
a driver-by-driver results of each race
}
\keyword{datasets}
