% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{safety_cars}
\alias{safety_cars}
\title{Safety Cars}
\format{
a data frame with four columns:
\itemize{
\item year (numeric)
\item name (character)
\item count (numeric)
\item laps (numeric)
}
}
\source{
https://www.reddit.com/r/formula1/comments/iv5683/comment/g5ppx4h/

https://f1.fandom.com/wiki/Safety_Car#Safety_car_deployments_in_Formula_One_races

fastf1 API
}
\usage{
safety_cars
}
\description{
contains data on safety cars for each grand prix, including
the total number of safety cars used and the number of laps under safety
car conditions.
}
\keyword{datasets}
